+++ CHANGELOG.md
@@ -6,3 +6,4 @@ v0.8.6
 - [REFACTOR] selectAll Return type in the DataController -> return List of Specific class objects
 - [TASK] change budget colors
 - [TASK] add total budget progressbar
+- [TASK] add sum of total used budget